adds condition to install script, updates rascaline install script

+++ demos/ensemble-deltamu/install_n2p2.sh
@@ -1,13 +1,28 @@
 #!/bin/bash
 
-echo "Cleaning conda cache"
-export LD_LIBRARY_PATH=
+# Check if the miniconda3 directory already exists
+if [ ! -d "./miniconda3" ]; then
+    echo "Miniconda is not installed. Installing Miniconda..."
 
-mkdir -p ./miniconda3
-wget https://repo.anaconda.com/miniconda/Miniconda3-latest-Linux-x86_64.sh -O ./miniconda3/miniconda.sh
-chmod u+rx ./miniconda3/miniconda.sh
-bash ./miniconda3/miniconda.sh -b -u -p ./miniconda3
-rm -rf ./miniconda3/miniconda.sh
+    # Create directory for Miniconda
+    mkdir -p ./miniconda3
+
+    # Download Miniconda
+    wget https://repo.anaconda.com/miniconda/Miniconda3-latest-Linux-x86_64.sh -O ./miniconda3/miniconda.sh
+
+    # Make the script executable
+    chmod u+rx ./miniconda3/miniconda.sh
+
+    # Execute the installer
+    bash ./miniconda3/miniconda.sh -b -u -p ./miniconda3
+
+    # Remove the installer script
+    rm -rf ./miniconda3/miniconda.sh
+
+    echo "Miniconda installed successfully."
+else
+    echo "Miniconda is already installed. Skipping installation."
+fi
 
 ./miniconda3/bin/conda create -n n2p2_install -y
